+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/AC_LookAtCamera.cs
@@ -1,6 +1,3 @@
-
-
-using Newtonsoft.Json;
 using UnityEngine;
 
 public class AC_LookAtCamera : AC_ConfigableUpdateComponentBase<Transform, AC_SOLookAtCameraConfig, AC_LookAtCamera.ConfigInfo>