+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/SO/AC_SOFixedRotationConfig.cs
@@ -2,5 +2,5 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "FixedRotation", fileName = "FixedRotationConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Transform/FixedRotation", fileName = "FixedRotationConfig")]
 public class AC_SOFixedRotationConfig : AC_SOConfigBase<AC_FixedRotation.ConfigInfo>{}