+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/SO/AC_SOLookAtCameraConfig.cs
@@ -2,5 +2,5 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "LookAtCamera", fileName = "LookAtCameraConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Transform/LookAtCamera", fileName = "LookAtCameraConfig")]
 public class AC_SOLookAtCameraConfig : AC_SOConfigBase<AC_LookAtCamera.ConfigInfo>{}