+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/SO/AC_SORotateByTargetMovementConfig.cs
@@ -2,5 +2,5 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "RotateByTargetMovement", fileName = "RotateByTargetMovementConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Transform/RotateByTargetMovement", fileName = "RotateByTargetMovementConfig")]
 public class AC_SORotateByTargetMovementConfig : AC_SOConfigBase<AC_RotateByTargetMovement.ConfigInfo> { }