+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/SO/AC_SOTiltByTargetMovementConfig.cs
@@ -2,5 +2,5 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "TiltByTargetMovement", fileName = "TiltByTargetMovementConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Transform/TiltByTargetMovement", fileName = "TiltByTargetMovementConfig")]
 public class AC_SOTiltByTargetMovementConfig : AC_SOConfigBase<AC_TiltByTargetMovement.ConfigInfo>{}