+++ Threeyes/SDK/Scripts/Component/Feature/AudioVisualizer/SO/AC_SOAudioVisualizer_IcoSphereConfig.cs
@@ -2,5 +2,5 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "AudioVisualizer_IcoSphere", fileName = "AudioVisualizer_IcoSphereConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Feature/AudioVisualizer/IcoSphere", fileName = "AudioVisualizer_IcoSphereConfig")]
 public class AC_SOAudioVisualizer_IcoSphereConfig : AC_SOConfigBase<AC_AudioVisualizer_IcoSphere.ConfigInfo> { }