+++ Threeyes/SDK/Scripts/Component/Feature/Clock/SO/AC_SOClockControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Feature/ClockController", fileName = "ClockControllerConfig")]
+public class AC_SOClockControllerConfig
+	: AC_SOConfigBase<AC_ClockController.ConfigInfo> { }