+++ Threeyes/SDK/Scripts/Component/Feature/Liquid/SO/AC_SOLiquidControllerConfig.cs
@@ -2,5 +2,5 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "LiquidController", fileName = "LiquidControllerConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Feature/LiquidController", fileName = "LiquidControllerConfig")]
 public class AC_SOLiquidControllerConfig : AC_SOConfigBase<AC_LiquidController.ConfigInfo>{}