+++ Threeyes/SDK/Scripts/Component/Feature/ObjectGenerator/SO/AC_SOObjectGeneratorConfig.cs
@@ -1,5 +1,5 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "ObjectGenerator", fileName = "ObjectGeneratorConfig")]
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Feature/ObjectGenerator", fileName = "ObjectGeneratorConfig")]
 public class AC_SOObjectGeneratorConfig : AC_SOConfigBase<AC_ObjectGenerator.ConfigInfo> { }