+++ Threeyes/SDK/Scripts/Component/Zibra Liquids/SO/AC_SOZibraLiquidAdvancedRenderParametersControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "ZibraLiquid/AdvancedRenderParametersController", fileName = "ZibraLiquidAdvancedRenderParametersControllerConfig")]
+public class AC_SOZibraLiquidAdvancedRenderParametersControllerConfig : AC_SOConfigBase<AC_ZibraLiquidAdvancedRenderParametersController.ConfigInfo> { }