+++ Threeyes/SDK/Scripts/Component/Zibra Liquids/SO/AC_SOZibraLiquidControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "ZibraLiquid/Controller", fileName = "ZibraLiquidControllerConfig")]
+public class AC_SOZibraLiquidControllerConfig : AC_SOConfigBase<AC_ZibraLiquidController.ConfigInfo> { }