+++ Threeyes/SDK/Scripts/Component/Zibra Liquids/SO/AC_SOZibraLiquidForceFieldControllerConfig.cs
@@ -1,0 +1,5 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "ZibraLiquid/ForceFieldController", fileName = "ZibraLiquidForceFieldControllerConfig")]
+public class AC_SOZibraLiquidForceFieldControllerConfig : AC_SOConfigBase<AC_ZibraLiquidForceFieldController.ConfigInfo> { }