+++ Threeyes/SDK/Scripts/Component/Zibra Liquids/SO/AC_SOZibraLiquidMaterialParametersControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "ZibraLiquid/MaterialParametersController", fileName = "ZibraLiquidMaterialParametersControllerConfig")]
+public class AC_SOZibraLiquidMaterialParametersControllerConfig : AC_SOConfigBase<AC_ZibraLiquidMaterialParametersController.ConfigInfo> { }