+++ Threeyes/SDK/Scripts/Component/Zibra Liquids/SO/AC_SOZibraLiquidSolverParametersControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "ZibraLiquid/SolverParametersController", fileName = "ZibraLiquidSolverParametersControllerConfig")]
+public class AC_SOZibraLiquidSolverParametersControllerConfig : AC_SOConfigBase<AC_ZibraLiquidSolverParametersController.ConfigInfo> { }